fix(core): set group message reaction seq should be Long

+++ acidify-core/src/commonMain/kotlin/org/ntqqrev/acidify/Bot.kt
@@ -1087,7 +1087,7 @@ class Bot(
      */
     suspend fun setGroupMessageReaction(
         groupUin: Long,
-        sequence: Int,
+        sequence: Long,
         code: String,
         isAdd: Boolean = true
     ) = client.callService(
